requirements-dev.txt
black==22.3.0
flake8==4.0.1
isort==5.10.1
pytest==7.1.2
pytest-cov==3.0.0
timm==0.5.4
torchvision==0.12.0